requirements.txt
alembic>=0.6.4
anyjson>=0.3.3
argparse
croniter>=0.3.4 # MIT License
eventlet>=0.13.0
happybase>=0.5,!=0.7
iso8601>=0.1.9
jsonpath-rw>=1.2.0,<2.0
jsonschema>=2.0.0,<3.0.0
keystonemiddleware>=1.0.0
lockfile>=0.8
lxml>=2.3
msgpack-python>=0.4.0
netaddr>=0.7.6
oslo.config>=1.2.1
oslo.vmware>=0.4                        # Apache-2.0
PasteDeploy>=1.5.0
pbr>=0.6,!=0.7,<1.0
pecan>=0.5.0
posix_ipc
oslo.messaging>=1.3.0
pysnmp>=4.2.1,<5.0.0
python-ceilometerclient>=1.0.6
python-glanceclient>=0.13.1
python-keystoneclient>=0.9.0
python-neutronclient>=2.3.5,<3
python-novaclient>=2.17.0
python-swiftclient>=2.0.2
pytz>=2010h
PyYAML>=3.1.0
requests>=1.1
six>=1.7.0
SQLAlchemy>=0.8.4,!=0.9.5,<=0.9.99
sqlalchemy-migrate>=0.9.1
stevedore>=0.14
tooz>=0.2
WebOb>=1.2.3
WSME>=0.6
